EOS-16522: sspl_setup_init.sh to python (added SetupError class)
